exercise 4.21 complete

+++ part4/bloglist/controllers/blogs.js
@@ -35,6 +35,17 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  if (decodedToken.id !== blogToDelete.user.toString()) { 
+    return response.status(401).json({ error: 'blog can only be edited by its creator' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
